requirements.txt
pytest==7.4.2
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
selenium==4.14.0
webdriver-manager==4.0.1
pytest-playwright==0.4.3
httpx==0.25.0 